+++ src/aula3/3CallerOtherContract.sol
@@ -4,17 +4,17 @@ pragma solidity ^0.8.20;
 //formas de se chamar outros contratos e modificar variaveis e enviar valores
 
 contract Callee {
-    uint public x;
-    uint public value;
+    uint256 public x;
+    uint256 public value;
 
-//setar valor de x na variavel global
-    function setX(uint _x) public returns (uint) {
+    //setar valor de x na variavel global
+    function setX(uint256 _x) public returns (uint256) {
         x = _x;
         return x;
     }
 
-//enviar valor de x e enviar Ether
-    function setXandSendEther(uint _x) public payable returns (uint, uint) {
+    //enviar valor de x e enviar Ether
+    function setXandSendEther(uint256 _x) public payable returns (uint256, uint256) {
         x = _x;
         value = msg.value;
 
@@ -22,19 +22,18 @@ contract Callee {
     }
 }
 
-
 //novo contrato que vamos chamar as funcoes do outro contrato e setar informacoes
 contract CallerFunctions {
-    function setX(Callee _callee, uint _x) public {
-        uint x = _callee.setX(_x);
+    function setX(Callee _callee, uint256 _x) public {
+        uint256 x = _callee.setX(_x);
     }
 
-    function setXFromAddress(address _addr, uint _x) public {
+    function setXFromAddress(address _addr, uint256 _x) public {
         Callee callee = Callee(_addr);
         callee.setX(_x);
     }
 
-    function setXandSendEther(Callee _callee, uint _x) public payable {
-        (uint x, uint value) = _callee.setXandSendEther{value: msg.value}(_x);
+    function setXandSendEther(Callee _callee, uint256 _x) public payable {
+        (uint256 x, uint256 value) = _callee.setXandSendEther{ value: msg.value }(_x);
     }
 }
