+++ src/aula5/1structs.sol
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.20;
 
-
 contract Todos {
     struct Todo {
         string text;
@@ -17,7 +16,7 @@ contract Todos {
         todos.push(Todo(_text, false));
 
         // key value mapping
-        todos.push(Todo({text: _text, completed: false}));
+        todos.push(Todo({ text: _text, completed: false }));
 
         // initialize an empty struct and then update it
         Todo memory todo;
@@ -29,19 +28,19 @@ contract Todos {
 
     // Solidity automatically created a getter for 'todos' so
     // you don't actually need this function.
-    function get(uint _index) public view returns (string memory text, bool completed) {
+    function get(uint256 _index) public view returns (string memory text, bool completed) {
         Todo storage todo = todos[_index];
         return (todo.text, todo.completed);
     }
 
     // update text
-    function updateText(uint _index, string calldata _text) public {
+    function updateText(uint256 _index, string calldata _text) public {
         Todo storage todo = todos[_index];
         todo.text = _text;
     }
 
     // update completed
-    function toggleCompleted(uint _index) public {
+    function toggleCompleted(uint256 _index) public {
         Todo storage todo = todos[_index];
         todo.completed = !todo.completed;
     }
