+++ src/aula5/exerc/oracleRealstate.sol
@@ -1,20 +1,26 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.20;
 
-
 contract oracleRealstate {
-
     address public owner; //dono do contrato oráculo e incorporadora/contrutora
     string public model; //modelo do imóvel
-    uint public valueRealEstate; //valor do imóvel
-    address public broker; //endereço do corretor 
-    uint public immutable commission; //comissão do corretor
-    uint public timepay; // a data que é para pagar a compra
+    uint256 public valueRealEstate; //valor do imóvel
+    address public broker; //endereço do corretor
+    uint256 public immutable commission; //comissão do corretor
+    uint256 public timepay; // a data que é para pagar a compra
 
     address public RealestateAddr;
 
-    constructor(address _owner, string memory _model,
-    uint _valueRealEstate, address _broker, uint _commission, uint _timepay) payable {
+    constructor(
+        address _owner,
+        string memory _model,
+        uint256 _valueRealEstate,
+        address _broker,
+        uint256 _commission,
+        uint256 _timepay
+    )
+        payable
+    {
         owner = _owner;
         model = _model;
         broker = _broker;
@@ -25,15 +31,12 @@ contract oracleRealstate {
         RealestateAddr = address(this);
     }
 
-    function newTimepay(uint _newtimepay) external onlyOwner{
+    function newTimepay(uint256 _newtimepay) external onlyOwner {
         timepay = _newtimepay;
     }
 
-    modifier onlyOwner(){
+    modifier onlyOwner() {
         require(msg.sender == owner, "not owner");
         _;
     }
-
-
-   
 }
